refactor: extract editor chrome controls
